refactor: renamed bottom bar icon
