+++ src/navigation/RootNavigator/HomeTabs/components/BottomBarIcon/index.tsx
